fix: make the dashboard work when opened through a tunnel
Three separate bugs made a remote dashboard unusable. All of them only
show up when the UI is not served from the machine running the server.

1. serverOrigin() built http://<host>:6869 from the page hostname. A
   tunnel only publishes the web port, so that host never answers, and
   an https page calling http is blocked as mixed content anyway. Ten
   call sites use it: asset/sfx/music upload, style logo and font,
   image background, voice cloning, translate-video source, skill
   generation. uploadOrigin() already solved this but only the mobile
   /m page called it, so the fix now lives in serverOrigin() and
   uploadOrigin() delegates to it.

2. POST /api/assets required a QR upload-session token for every
   non-loopback request. Through a tunnel the desktop dashboard is not
   loopback, so it got 403 UPLOAD_TOKEN_INVALID even while holding
   AIEV_API_TOKEN, which already permits running the agent and deleting
   projects. Requests carrying that token are now exempt. The QR flow is
   untouched: a phone has no master token and still needs a valid ?k=.

3. The Next proxy caps request bodies at 10MB. With uploads routed
   through it, anything larger was truncated, multer waited for bytes
   that never came, the 45s stall guard aborted the request and the user
   saw HTTP 500. proxyClientMaxBodySize now matches multer's own 2GB
   limit so neither side silently caps the other.

Measured on a Cloudflare Tunnel: 50MB upload returns 201 in 3.7s;
150MB returns 413 from Cloudflare itself, which is the platform's
100MB free-plan cap and outside what this repo can fix.

Also moves cookieValue into util.ts, where the new hasMasterToken needs
it, instead of keeping a second copy in index.ts.
